+++ demo/gulpfile.js
@@ -21,7 +21,7 @@ var paths = {
 gulp.task('default', ['watch']);
 
 gulp.task('ng2Component', function() {
-    var dest = plugins.util.env.module ? 'src/app/'+plugins.util.env.module+'/'+plugins.util.env.name+'/': 'src/app/'+plugins.util.env.name+'/';
+    var dest = plugins.util.env.module ? '../src/app/'+plugins.util.env.module+'/'+plugins.util.env.name+'/': '../src/app/'+plugins.util.env.name+'/';
     return gulp.src(paths.gulp.templates + 'view.tpl.html')
     .pipe(plugins.plumber())
     .pipe(plugins.template({name: plugins.util.env.name}))
@@ -55,38 +55,38 @@ gulp.task('ng2Module', function() {
     .pipe(plugins.plumber())
     .pipe(plugins.template({name: plugins.util.env.name}))
     .pipe(plugins.rename({ basename: plugins.util.env.name, extname: '.component.html'}))
-    .pipe(gulp.dest('demo/src/app/'+ plugins.util.env.name + '/'))
+    .pipe(gulp.dest('../src/app/'+ plugins.util.env.name + '/'))
      //SCSS
     
     .pipe(plugins.addSrc.append(paths.gulp.templates + 'component.tpl.scss'), { passthrough: true })
     .pipe(plugins.template({name: capitalize(plugins.util.env.name)}))
     .pipe(plugins.rename({ basename: plugins.util.env.name, extname: '.component.scss'}))
-    .pipe(gulp.dest('demo/src/app/'+ plugins.util.env.name + '/'))
+    .pipe(gulp.dest('../src/app/'+ plugins.util.env.name + '/'))
 
     //Component
     
     .pipe(plugins.addSrc.append(paths.gulp.templates + 'component.tpl.ts'), { passthrough: true })
     .pipe(plugins.template({name: plugins.util.env.name, capitalname: capitalize(plugins.util.env.name)}))
     .pipe(plugins.rename({ basename: plugins.util.env.name, extname: '.component.ts'}))
-    .pipe(gulp.dest('demo/src/app/'+ plugins.util.env.name + '/'))
+    .pipe(gulp.dest('../src/app/'+ plugins.util.env.name + '/'))
 
     //Module
     .pipe(plugins.addSrc.append(paths.gulp.templates + 'module.tpl.ts'), { passthrough: true })
     .pipe(plugins.template({name: plugins.util.env.name, capitalname: capitalize(plugins.util.env.name)}))
     .pipe(plugins.rename({ basename: plugins.util.env.name, extname: '.module.ts'}))
-    .pipe(gulp.dest('demo/src/app/'+ plugins.util.env.name + '/'))
+    .pipe(gulp.dest('../src/app/'+ plugins.util.env.name + '/'))
    
     //Routes
     
     .pipe(plugins.addSrc.append(paths.gulp.templates + 'routes.tpl.ts'), { passthrough: true })
     .pipe(plugins.template({name: plugins.util.env.name, capitalname: capitalize(plugins.util.env.name)}))
     .pipe(plugins.rename({ basename: plugins.util.env.name, extname: '.routes.ts'}))
-    .pipe(gulp.dest('demo/src/app/'+ plugins.util.env.name + '/'))
+    .pipe(gulp.dest('../src/app/'+ plugins.util.env.name + '/'))
 
 });
 
 gulp.task('ng2Service', function() {
-    var dest = 'src/app/'+plugins.util.env.module+'/';
+    var dest = '../src/app/'+plugins.util.env.module+'/';
     return gulp.src(paths.gulp.templates + 'factory.tpl.js')
     .pipe(plugins.template({name: plugins.util.env.name, module:plugins.util.env.module}))
     .pipe(plugins.rename({ basename: plugins.util.env.name, extname: '.service.js'}))
@@ -95,7 +95,7 @@ gulp.task('ng2Service', function() {
 
 
 gulp.task('ngDirective', function() {
-    var dest = 'src/app/'+plugins.util.env.module+'/';
+    var dest = '../src/app/'+plugins.util.env.module+'/';
 
     return gulp.src(paths.gulp.templates + 'directive.tpl.js')
     .pipe(plugins.template({name: plugins.util.env.name, restrict: plugins.util.env.restrict}))
@@ -103,7 +103,7 @@ gulp.task('ngDirective', function() {
     .pipe(gulp.dest(dest));
 });
 gulp.task('ngComponent', function() {
-    var dest = 'src/app/'+plugins.util.env.module+'/';
+    var dest = '../src/app/'+plugins.util.env.module+'/';
 
     return gulp.src(paths.gulp.templates + 'component.tpl.js')
     .pipe(plugins.template({name: plugins.util.env.name, restrict: plugins.util.env.restrict}))
@@ -111,7 +111,7 @@ gulp.task('ngComponent', function() {
     .pipe(gulp.dest(dest));
 });
 gulp.task('ngFilter', function() {
-    var dest = 'src/app/'+plugins.util.env.module+'/';
+    var dest = '../src/app/'+plugins.util.env.module+'/';
 
     return gulp.src(paths.gulp.templates + 'filter.tpl.js')
     .pipe(plugins.template({name: plugins.util.env.name}))