+++ demo/src/app/core/core.module.ts
@@ -6,15 +6,27 @@ import { FormsModule } from '@angular/forms';
 import { NumberFormatterPipe } from './misc/number-formatter.pipe';
 
 import { FormatService } from './misc/format.service';
+import { NavbarModule } from './navbar/navbar.module';
+
 import { HomeModule } from '../home/home.module';
 import { CardsActivityModule } from '../cards/cards.module';
+import { ListsActivityModule } from './../lists/lists.module';
+import { FormsActivityModule } from '../forms/forms.module';
+import { GridActivityModule } from './../grid/grid.module';
+import { LayoutModule } from './../layout/layout.module';
+import { NavService } from './navmenu/nav.service';
 
 @NgModule({
   imports: [
     CommonModule,
     FormsModule,
+    NavbarModule,
     HomeModule,
-    CardsActivityModule
+    CardsActivityModule,
+    ListsActivityModule,
+    FormsActivityModule,
+    GridActivityModule,
+    LayoutModule
 
   ],
   declarations: [
@@ -23,13 +35,19 @@ import { CardsActivityModule } from '../cards/cards.module';
 
   providers: [
     FormatService,
+    NavService
   ],
   exports: [
     CommonModule,
     FormsModule,
+    NavbarModule,
     HomeModule,
     CardsActivityModule,
+    ListsActivityModule,
+    FormsActivityModule,
     NumberFormatterPipe,
+    GridActivityModule,
+    LayoutModule
   ]
 })
 export class CoreModule { }