+++ demo/src/app/forms/forms.routes.ts
@@ -1,0 +1,11 @@
+import { Routes } from '@angular/router';
+import { DataResolver } from '../app.resolver';
+import { FormsComponent } from './forms.component';
+
+export const ROUTES: Routes = [
+    {
+         path: 'forms',
+         component: FormsComponent,
+         data: { title: 'Forms' }
+    }
+];