+++ demo/src/app/home/home.component.ts
@@ -1,3 +1,4 @@
+
 import {
     Component,
     OnInit,
@@ -6,6 +7,7 @@ import {
 import {
     ToastyService
 } from 'ng2-toasty';
+import { Header } from './../../../../src/header/header';
 
 @Component({
     encapsulation: ViewEncapsulation.None,
@@ -34,10 +36,41 @@ export class HomeComponent implements OnInit {
         {
             label: 'Lists',
             illustration: '/assets/svg/lists-illustration.svg',
-            comming: true,
+            comming: false,
+            slug: 'lists',
             msg: `Lists present multiple line items vertically as a single continuous element`
+        },
+        {
+            label: 'Forms',
+            illustration: '/assets/svg/forms-illustration.svg',
+            comming: false,
+            slug: 'forms',
+            msg: `Forms represent a way for the user to interact with the business logic`
+        },
+        {
+            label: 'Grid',
+            illustration: '/assets/svg/grid-illustration.svg',
+            comming: false,
+            slug: 'grid',
+            // tslint:disable-next-line:max-line-length
+            msg: `A grid list consists of a repeated pattern of cells arrayed in a vertical and horizontal layout.`
+        },
+        {
+            label: 'Layout',
+            illustration: '/assets/svg/layout-illustration.svg',
+            comming: false,
+            slug: 'layout',
+            // tslint:disable-next-line:max-line-length
+            msg: `A layout list consists of a repeated pattern of cells arrayed in a vertical and horizontal layout.`
         }
     ];
+    public header: Header = new Header(
+        {
+            bg: '../assets/img/header-canaima.png',
+            customClass: 'header-home',
+            parallax: true
+        }
+    );
     constructor(
         public toastService: ToastyService
     ) {}