+++ demo/src/app/layout/layout.routes.ts
@@ -1,0 +1,11 @@
+import { Routes } from '@angular/router';
+import { DataResolver } from '../app.resolver';
+import { LayoutComponent } from './layout.component';
+
+export const ROUTES: Routes = [
+    {
+         path: 'layout',
+         component: LayoutComponent,
+         data: { title: 'Layout' }
+    }
+];