+++ demo/src/app/lists/lists.routes.ts
@@ -1,0 +1,12 @@
+import { Routes } from '@angular/router';
+
+import { DataResolver } from '../app.resolver';
+import { ListsComponent } from './lists.component';
+
+export const ROUTES: Routes = [
+    {
+         path: 'lists',
+         component: ListsComponent,
+         data: { title: 'Lists' }
+    }
+];