+++ demo/src/app/shared/shared.module.ts
@@ -8,6 +8,7 @@ import {
   RequestOptions,
   Http
 } from '@angular/http';
+import { HttpClient } from '@angular/common/http';
 import {
   InterceptorService,
   Interceptor,
@@ -21,22 +22,22 @@ import {
 import {
   TranslateHttpLoader
 } from '@ngx-translate/http-loader';
-import { NavbarModule } from './navbar/navbar.module';
 
-import { NgxCanaimaModule } from './../../../../src/index';
+import {
+  NgxCanaimaModule
+} from './../../../../src/index';
 
 @NgModule({
   imports: [
     CommonModule,
     FormsModule,
     FlexLayoutModule,
-    NavbarModule,
     NgxCanaimaModule,
     TranslateModule.forRoot({
       loader: {
         provide: TranslateLoader,
-        useFactory: (http: Http) => new TranslateHttpLoader(http, '/assets/i18n/', '.json'),
-        deps: [Http]
+        useFactory: (http: HttpClient) => new TranslateHttpLoader(http, '/assets/i18n/', '.json'),
+        deps: [HttpClient]
       }
     })
   ],
@@ -50,7 +51,6 @@ import { NgxCanaimaModule } from './../../../../src/index';
     FormsModule,
     FlexLayoutModule,
     HttpModule,
-    NavbarModule,
     NgxCanaimaModule
   ]
 })