+++ src/app/layout/layout.module.ts
@@ -1,0 +1,21 @@
+import { NgModule } from '@angular/core';
+import {
+  RouterModule,
+  PreloadAllModules
+} from '@angular/router';
+import { ROUTES } from './layout.routes';
+import { LayoutComponent } from './layout.component';
+
+import { SharedModule } from '../shared/shared.module';
+@NgModule({
+    imports: [
+        SharedModule,
+        RouterModule.forRoot(ROUTES, { useHash: true, preloadingStrategy: PreloadAllModules })
+    ],
+    exports: [],
+    declarations: [
+        LayoutComponent
+    ],
+    providers: [],
+})
+export class LayoutModule { }