+++ src/index.ts
@@ -1,20 +1,55 @@
-
 import { NgModule, ModuleWithProviders } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { CardsModule } from './cards/cards.module';
+import { ListsModule } from './lists/lists.module';
+import { GridListsModule } from './grid-list/grid-list.module';
 
-export * from './cards/card/card.component';
+import { DynamicFormsModule } from './forms/dynamic-forms/dynamic-forms.module';
+import { ToolbarModule } from './toolbar/toolbar.module';
+import { HeaderModule } from './header/header.module';
+import { NavmenuModule } from './navmenu/navmenu.module';
+import { SegmentModule } from './segment/segment.module';
+import { ButtonsModule } from './buttons/buttons.module';
+
 export * from './action';
+export * from './action-dropdown';
+export * from './bg-image-preview.directive';
+export * from './image-preview.directive';
+export * from './cards';
+export * from './forms';
+export * from './lists';
+export * from './grid-list';
+export * from './segment';
+export * from './navmenu';
+export * from './toolbar';
+export * from './header';
 
 @NgModule({
   imports: [
     CommonModule,
-    CardsModule
+    CardsModule,
+    ListsModule,
+    DynamicFormsModule,
+    GridListsModule,
+    ToolbarModule,
+    NavmenuModule,
+    HeaderModule,
+    SegmentModule,
+    ButtonsModule
   ],
   declarations: [
+
   ],
   exports: [
     CardsModule,
+    ListsModule,
+    DynamicFormsModule,
+    GridListsModule,
+    ToolbarModule,
+    NavmenuModule,
+    HeaderModule,
+    SegmentModule,
+    ButtonsModule
   ]
 })
 export class NgxCanaimaModule {