+++ src/lists/list-item-check/list-item-check.component.ts
@@ -1,0 +1,34 @@
+
+import {
+    Component,
+    OnInit,
+    OnChanges,
+    Input,
+    Output,
+    EventEmitter,
+    ViewChild
+} from '@angular/core';
+import { Action } from './../../action';
+@Component({
+    selector: 'as-list-item-check',
+    templateUrl: 'list-item-check.component.html'
+})
+export class ListItemCheckComponent {
+    @Input('index') public index: string;
+    @Input('title') public title: string;
+    @Input('subtitle') public subtitle: string;
+    @Input('hasLine') public hasLine: boolean;
+    @Input('least') public least: string;
+    @Input('actions') public actions: [Action];
+    @Input('customClass') public customClass: string;
+
+    @Output() public onSelected = new EventEmitter<any>();
+    public isChecked = false;
+    /**
+     * select
+     */
+    public select() {
+        this.onSelected.emit(this.index);
+        this.isChecked = !this.isChecked;
+    }
+}