+++ src/lists/list-item-check/list-item-check.ts
@@ -1,0 +1,11 @@
+import { Action } from './../../action';
+import { ListItemBase } from './../list-item-base';
+export class ListItemCheck extends ListItemBase {
+    public least?: string;
+    constructor(
+        options: {} = {}
+    ) {
+        super(options);
+        this.least = options['least'] || null;
+    }
+}