+++ CHANGELOG.md
@@ -1,3 +1,39 @@
+## 0.4.1 - 2017-09-13
+### Fixed
+- Export Action Dropdown
+- Export Image and Bg Preview Directives
+## 0.4.0 - 2017-09-12
+### Added
+- Toolbar Component
+- Header Component
+- Parallax Header Component
+- Back Button Component
+- Segments Component
+- Button Directive for size, colors.
+- Button Icon Directive
+- Loading Button
+### Changes
+- Distribution of sass files to mixins, variables, effects etc.
+- Material theme
+## 0.3.0 - 2017-09-02
+### Added
+- Dynamic Forms
+- Text Input
+- Select Input
+- Radio Input
+- Checkbox Input
+- TextArea Input
+- List Module
+- Basic List Item
+- Avatar List Item
+- Icon List Item
+- Checkbox List Item
+- File Qeue List item
+- Property (key, value) List Item
+- Grid List Module with responsive
+### Fixed
+- Card Image transclude content
+- Cards visual validations
 ## 0.2.0 - 2017-04-25
 ### Added
 - Image Card Component